style: 코드 스타일 정리
