+++ src/utils/scrollFunction.ts
@@ -4,8 +4,8 @@
  */
 
 export const moveToBottom = () => {
-    window.scrollTo({ 
-      top: document.body.scrollHeight,
-      behavior: "smooth" 
-    });
-}
+  window.scrollTo({
+    top: document.body.scrollHeight,
+    behavior: 'smooth',
+  });
+};
